+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## Unreleased
+- added fail on missing service flag
+- ignore checks during initialization flag
 
 ## [0.0.5] - 2015-11-26
 ### Changed